Move logic to service - paciente

+++ src/main/java/com/backend/clinica/service/PacienteService.java
@@ -1,15 +1,18 @@
 package com.backend.clinica.service;
 
 import com.backend.clinica.repository.PacienteRepository;
+import com.backend.clinica.exception.EntityAlreadyExistsException;
+import com.backend.clinica.exception.EntityNotFoundException;
 import com.backend.clinica.entity.PacienteModel;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
-public class PacienteService implements IService<PacienteModel> {
+public class PacienteService implements IService2<PacienteModel> {
     @Autowired
     private PacienteRepository pacienteRepository;
 
@@ -19,26 +22,61 @@ public class PacienteService implements IService<PacienteModel> {
     }
 
     @Override
-    public Optional<PacienteModel> findById(Long id) {
-        return pacienteRepository.findById(id);
+    public PacienteModel findById(Long id) throws EntityNotFoundException {
+        Optional<PacienteModel> paciente = pacienteRepository.findById(id);
+
+        if (paciente.isPresent()) {
+            return paciente.get();
+        } else {
+            throw new EntityNotFoundException("Paciente", "id", id);
+        }
     }
 
     @Override
-    public PacienteModel create(PacienteModel pacienteModel) {
+    public PacienteModel create(PacienteModel pacienteModel) throws EntityAlreadyExistsException {
+        Optional<PacienteModel> pacienteEqualsByDni = pacienteRepository.findByDni(pacienteModel.getDni());
+
+        if (pacienteEqualsByDni.isPresent()) {
+            throw new EntityAlreadyExistsException("Paciente", "dni", pacienteModel.getDni());
+        } else {
+            return pacienteRepository.save(pacienteModel);
+        }
+    }
+
+    @Override
+    public PacienteModel update(PacienteModel pacienteModel) throws EntityNotFoundException, EntityAlreadyExistsException {
+        Optional<PacienteModel> pacienteFindById = pacienteRepository.findById(pacienteModel.getPacienteID());
+        if (pacienteFindById.isEmpty()) {
+            throw new EntityNotFoundException("Paciente", "id", pacienteModel.getPacienteID());
+        }
+
+        Optional<PacienteModel> pacienteEqualsByDni = pacienteRepository.findByDni(pacienteModel.getDni());
+        if (pacienteEqualsByDni.isPresent() && !Objects.equals(pacienteEqualsByDni.get().getPacienteID(), pacienteModel.getPacienteID())) {
+            throw new EntityAlreadyExistsException("Paciente", "dni", pacienteModel.getDni());
+        }
+
         return pacienteRepository.save(pacienteModel);
     }
 
     @Override
-    public void update(PacienteModel pacienteModel) {
-        pacienteRepository.save(pacienteModel);
+    public PacienteModel delete(Long id) throws EntityNotFoundException {
+        Optional<PacienteModel> paciente = pacienteRepository.findById(id);
+
+        if (paciente.isPresent()) {
+            pacienteRepository.deleteById(id);
+            return paciente.get();
+        } else {
+            throw new EntityNotFoundException("Paciente", "id", id);
+        }
     }
 
-    @Override
-    public void delete(Long id) {
-        pacienteRepository.deleteById(id);
-    }
+    public PacienteModel findByDni(String dni) throws EntityNotFoundException {
+        Optional<PacienteModel> paciente = pacienteRepository.findByDni(dni);
 
-    public Optional<PacienteModel> findByDni(String dni) {
-        return pacienteRepository.findByDni(dni);
+        if (paciente.isPresent()) {
+            return paciente.get();
+        } else {
+            throw new EntityNotFoundException("Paciente", "dni", dni);
+        }
     }
 }
